Moved some test functions. Renamed a file.
