+++ Testing/testSendReceiveCommand.cxx
@@ -11,59 +11,14 @@
 #include <vtkImageDifference.h>
 #include "IGTLIOFixture.h"
 #include "igtlioSession.h"
-
-bool compareID(igtlio::CommandDevicePointer a, igtlio::CommandDevicePointer b)
-{
-  if (!a || !b)
-  {
-    std::cout << "FAILURE: empty device" << std::endl;
-    return false;
-  }
-
-  if (a->GetContent().id != b->GetContent().id)
-  {
-    std::cout << "FAILURE: Command IDs dont match." << std::endl;
-    return false;
-  }
-
-  return true;
-}
-
-bool compare(igtlio::CommandDevicePointer a, igtlio::CommandDevicePointer b)
-{
-  if (!a || !b)
-  {
-    std::cout << "FAILURE: empty device" << std::endl;
-    return false;
-  }
-
-  if (a->GetContent().name != b->GetContent().name)
-  {
-    std::cout << "FAILURE: Command names dont match." << std::endl;
-    return false;
-  }
-
-  if (a->GetContent().id != b->GetContent().id)
-  {
-    std::cout << "FAILURE: Command IDs dont match." << std::endl;
-    return false;
-  }
-
-  if (a->GetContent().content != b->GetContent().content)
-  {
-    std::cout << "FAILURE: Command content dont match." << std::endl;
-    return false;
-  }
-
-  return true;
-}
-
+#include "igtlioTestUtilities.h"
 
 ///
 /// Setup a client and server.
 /// Send a COMMAND from client to server.
 /// Server replies to the client with a RTS_COMMAND
 ///
+
 int main(int argc, char **argv)
 {
   ClientServerFixture fixture;
@@ -80,38 +35,47 @@ int main(int argc, char **argv)
   std::cout << "*** Connection done" << std::endl;
   //---------------------------------------------------------------------------
 
+  std::string device_name = "TestDevice";
   igtlio::CommandDevicePointer clientDevice;
-  clientDevice = fixture.Client.Session->SendCommandQuery("TestDevice_Command",
-                                                          "GetDeviceParameters",
+  clientDevice = fixture.Client.Session->SendCommandQuery(device_name,
+                                                          "Get",
                                                           "<Command>\n"
                                                           "  <Parameter Name=\"Depth\" />\n"
                                                           "</Command>",
                                                           igtlio::ASYNCHRONOUS);
+
   std::cout << "*** COMMAND query sent from Client" << std::endl;
   //---------------------------------------------------------------------------
 
   if (!fixture.LoopUntilEventDetected(&fixture.Server, igtlio::Logic::CommandQueryReceivedEvent))
     return 1;
+
   std::cout << "*** COMMAND query received by Server" << std::endl;
   //---------------------------------------------------------------------------
 
+  if(device_name != clientDevice->GetDeviceName())
+  {
+	  return 1;
+  }
+
   igtlio::CommandDevicePointer serverDevice;
   serverDevice = fixture.Server.Session->SendCommandResponse(clientDevice->GetDeviceName(),
-                                                             "GetDeviceParameters",
+                                                             "Get",
                                                              "<Command>\n"
-                                                             "  <Result>GetDeviceParameters: success</Result>\n"
-                                                             "  <Parameter Name=\"Depth\" Value=\"45\" />\n"
+                                                             "  <Result success=true> <Parameter Name=”Depth” /> </Result>\n"
                                                              "</Command>");
   if (!serverDevice)
   {
     std::cout << "FAILURE: Server did not send response." << std::endl;
     return false;
   }
+
   std::cout << "*** RTS_COMMAND response sent from Server" << std::endl;
   //---------------------------------------------------------------------------
 
   if (!fixture.LoopUntilEventDetected(&fixture.Client, igtlio::Logic::CommandResponseReceivedEvent))
     return 1;
+
   std::cout << "*** RTS_COMMAND response received by Client" << std::endl;
   //---------------------------------------------------------------------------
 
@@ -119,14 +83,14 @@ int main(int argc, char **argv)
   if (!clientDevice->GetQueries().empty())
     query = clientDevice->GetQueries()[0];
 
-  if (!compareID(igtlio::CommandDevice::SafeDownCast(query.Query),
+  if (!igtlio::compareID(igtlio::CommandDevice::SafeDownCast(query.Query),
                igtlio::CommandDevice::SafeDownCast(query.Response)))
   {
     std::cout << "FAILURE: Query and response dont match." << std::endl;
     return 1;
   }
 
-  if (!compare(serverDevice,
+  if (!igtlio::compare(serverDevice,
                igtlio::CommandDevice::SafeDownCast(query.Response)))
   {
     std::cout << "FAILURE: Received response not equal to what the Server sent." << std::endl;
